refactor: extract handleMainRequest from utility-bridge to stay under 800 lines
